refactor(gui): stage 6 — extract SessionTerm + focus pipeline; main.js is a composition root
- src/app/session-term.js: the SessionTerm class, font helpers, and
  the ensureTerm factory — verbatim. Imports view/sidebar/events/focus
  directly (none of them import it back; the reverse edges are the
  deps injected at wiring time).
- src/app/focus.js: setActive, the #186/#217 focus guard,
  setFocusedTile/applyFocus retry pipeline, focusSnapshot (now reading
  anyModalOpen()), debug checker, focusActiveTerm/refocusActiveTerm.
  ensureTerm is its only injected dep.
- src/app/trace.js: the scroll tracer singleton (was inline in main).
- main.js: 3,559 → 186 lines across the series — now only the palette
  command table, the init/deps wiring block, footer hints, the sidebar
  resizer, and bootstrap.

Suites: vitest 149, mock e2e 52(+1 skip), e2e-real 5.
