+++ cmd/hivegui/frontend/src/main.js
@@ -1,1114 +1,45 @@
+// Composition root. Every subsystem lives in src/app/* (pure logic in
+// src/lib/*); this file builds the command table, injects the
+// cross-module callbacks, and boots the control connection. Keep it
+// free of behavior — if a function body wants to live here, it almost
+// certainly belongs in a module.
+
 import '@xterm/xterm/css/xterm.css';
-import { Terminal } from '@xterm/xterm';
-import { FitAddon } from '@xterm/addon-fit';
-import { WebglAddon } from '@xterm/addon-webgl';
-import { WebLinksAddon } from '@xterm/addon-web-links';
 
 import {
-  ConnectControl, OpenSession, CloseAttach,
-  WriteStdin, ResizeSession, RequestScrollbackReplay,
-  CreateSession, DuplicateSession, KillSession, RestartSession, UpdateSession, ListAgents,
-  CreateProject, KillProject, UpdateProject,
-  LaunchDir, PickDirectory, OpenNewWindow, CloseWindow,
-  IsGitRepo, OpenURL, OpenTerminalAt, Notify, Confirm,
-  RestartDaemon, CheckForUpdate, SetClipboardText,
-  EventsOn, WindowSetTitle, ClipboardGetText,
+  ConnectControl, KillSession, OpenNewWindow, CloseWindow, OpenTerminalAt,
 } from './bridge.js';
-import { isMac, cmdOrCtrl } from './lib/platform.js';
-import { isShiftEnter, NEWLINE_SEQ } from './lib/keymap.js';
-import { computeGridDims, buildGridLayout, computeSpatialMove } from './lib/grid.js';
-import { DEFAULT_FONT_SIZE, clampFont } from './lib/font.js';
-import { VIEW_STORAGE_KEY } from './lib/view.js';
-import { filterMinimized } from './lib/minimized.js';
-import { snapVisibleTermsToBottom } from './lib/view-scroll.js';
+import { isMac } from './lib/platform.js';
+import { paletteShortcuts, footerHints } from './lib/shortcuts.js';
+import { state } from './app/state.js';
+import { setStatus, reportFailure } from './app/dom.js';
+import { activeCwd } from './app/selectors.js';
+import { scrollTrace } from './app/trace.js';
 import {
-  decideFocusAction, ACTION_CLEAR, ACTION_PRESERVE, ACTION_FOCUS,
-} from './lib/focus.js';
-import { readProjectId, readWorktreeBranch } from './lib/wire.js';
-import {
-  shouldRefreshOnVisibility,
-  recoverFromContextLoss,
-  bindDprWatcher,
-} from './lib/renderer-recovery.js';
-import {
-  shouldRequestReplay, REPLAY_DEBOUNCE_MS, handleScrollbackEvent,
-  applyRebaseline,
-} from './lib/scrollback.js';
-import { shortcutGroups, paletteShortcuts, footerHints } from './lib/shortcuts.js';
-import { emptyStateModel } from './lib/empty-state.js';
-import { pruneCollapsed } from './lib/collapsed.js';
-import { createScrollTrace } from './lib/scroll-debug.js';
-import { state, saveCollapsed } from './app/state.js';
-import { termsHost, projectsUL, setStatus, flashStatus, reportFailure } from './app/dom.js';
-import { orderedSessions, activeCwd, activeProjectId, resolveSessionCwd } from './app/selectors.js';
-import { anyModalOpen } from './app/modals/registry.js';
-import {
-  launcherEl, launcherState, moveLauncherSelection, activateLauncherSelection,
-  openLauncher, closeLauncher, duplicateActiveSession, restartActiveSession,
+  openLauncher, duplicateActiveSession, restartActiveSession,
   duplicateActiveSessionChooseTool, initLauncher,
 } from './app/modals/launcher.js';
-import { editorEl, openProjectEditor, initProjectEditor } from './app/modals/project-editor.js';
-import { openCommandPalette, initCommandPalette } from './app/modals/command-palette.js';
-import { openHelpOverlay, closeHelpOverlay, toggleHelpOverlay, initHelpOverlay } from './app/modals/help-overlay.js';
-import { renderSidebar, updateSidebarSelection, initSidebar } from './app/sidebar.js';
-import { onSessionBell, clearAttention, wireDaemonEvents } from './app/events.js';
+import { openProjectEditor, initProjectEditor } from './app/modals/project-editor.js';
+import { initCommandPalette } from './app/modals/command-palette.js';
+import { openHelpOverlay, initHelpOverlay } from './app/modals/help-overlay.js';
+import { initSidebar } from './app/sidebar.js';
+import { wireDaemonEvents } from './app/events.js';
 import { isDaemonRestarting, initBanners } from './app/banners.js';
 import {
-  showSingle, switchTo, switchToProject, updateAppTitle, renderGrid,
-  gridSpatialMove, shiftActiveProject, gridScopeSessions, minimizeSession,
-  restoreSession, renderMinimizedTray, renderEmptyState, setView, initView,
+  switchTo, switchToProject, updateAppTitle, renderMinimizedTray,
+  renderEmptyState, shiftActiveProject, initView,
 } from './app/view.js';
 import {
   initKeyboard, toggleSidebar, toggleProjectGrid, toggleAllGrid,
   confirmAndDeleteProject, deleteActiveProject, navSession, reorderActive,
   switchToNthSession,
 } from './app/keyboard.js';
-
-// Scroll/replay tracer — gated on localStorage hive.debug = '1' (the
-// focus consistency checker's switch). The gate is latched here at
-// module init: users hitting scroll-jump bugs set the key, RELOAD,
-// reproduce, then dump window.__hive_scrolltrace. The e2e-real scroll
-// specs arm it via addInitScript (before main.js runs) and read the
-// ring to prove a scenario actually fired replays.
-const scrollTrace = createScrollTrace({
-  enabled: (() => {
-    try { return localStorage.getItem('hive.debug') === '1'; } catch { return false; }
-  })(),
-});
-if (typeof window !== 'undefined') {
-  window.__hive_scrolltrace = scrollTrace.ring;
-}
-
-// ---------- session terminal ----------
-
-class SessionTerm {
-  constructor(info) {
-    this.info = info;
-    // Per-session UTF-8 decoder. Streaming mode buffers partial multi-byte
-    // sequences at chunk boundaries; sharing one decoder across sessions
-    // contaminates session B's bytes with session A's pending tail bytes,
-    // producing garbled glyphs — most visible with multi-byte-heavy output
-    // (emojis, box-drawing, Powerline glyphs from Claude, etc.).
-    this.decoder = new TextDecoder('utf-8', { fatal: false });
-    this.host = document.createElement('div');
-    this.host.className = 'term-host';
-    this.host.dataset.sid = info.id;
-    this.host.style.setProperty('--session-color', info.color || '#888');
-
-    // Tile header (only visible in grid mode via CSS).
-    this.header = document.createElement('div');
-    this.header.className = 'tile-header';
-    this.header.setAttribute('aria-label', `Session ${info.name}`);
-    this.tileColor = document.createElement('span');
-    this.tileColor.className = 'tile-color';
-    this.tileName = document.createElement('span');
-    this.tileName.className = 'tile-name';
-    this.tileName.textContent = info.name;
-    this.tileWorktree = document.createElement('span');
-    this.tileWorktree.className = 'worktree-glyph';
-    this.tileWorktree.textContent = '⎇';
-    {
-      const wtBranch = info.worktreeBranch ?? info.worktree_branch;
-      if (wtBranch) {
-        this.tileWorktree.title = `Worktree: ${wtBranch}`;
-      } else {
-        this.tileWorktree.style.display = 'none';
-      }
-    }
-    this.tileProject = document.createElement('span');
-    this.tileProject.className = 'tile-project';
-    // OSC-set window title from the running TUI (vim, htop, claude…).
-    // Sits between the session name and the project label so the
-    // user can tell at a glance what each tile is currently doing.
-    this.tileTermTitle = document.createElement('span');
-    this.tileTermTitle.className = 'tile-term-title';
-    this.tileMinimize = document.createElement('button');
-    this.tileMinimize.className = 'tile-minimize';
-    this.tileMinimize.type = 'button';
-    this.tileMinimize.title = 'Minimize (hide from grid)';
-    this.tileMinimize.setAttribute('aria-label', 'Minimize session');
-    this.tileMinimize.textContent = '–';
-    this.tileMinimize.addEventListener('mousedown', (e) => {
-      // Block the surrounding tile mousedown so minimizing doesn't
-      // also select / switch to this tile.
-      e.stopPropagation();
-    });
-    this.tileMinimize.addEventListener('click', (e) => {
-      e.stopPropagation();
-      minimizeSession(this.info.id);
-    });
-    this.header.append(this.tileColor, this.tileName, this.tileWorktree, this.tileTermTitle, this.tileProject, this.tileMinimize);
-
-    this.body = document.createElement('div');
-    this.body.className = 'term-body';
-
-    this.host.append(this.header, this.body);
-    document.getElementById('terms').appendChild(this.host);
-
-    this.term = new Terminal({
-      fontFamily: 'Menlo, "DejaVu Sans Mono", monospace',
-      fontSize: state.fontSize,
-      // cursorBlink causes a repaint twice a second per terminal —
-      // material on older machines with many tiles. Off by default.
-      cursorBlink: false,
-      scrollback: 5000,
-      smoothScrollDuration: 0,
-      theme: { background: '#000000' },
-      // Route OSC 8 hyperlinks (used by Claude CLI and others) through
-      // the OS default browser via the Wails backend.
-      linkHandler: {
-        activate: (_e, uri) => { if (uri) OpenURL(uri).catch(reportFailure('open link')); },
-      },
-    });
-    this.fit = new FitAddon();
-    this.term.loadAddon(this.fit);
-    this.term.open(this.body);
-
-    // Single source of truth for "tile geometry changed". Fires post-
-    // layout, only when the body's box actually resizes — covers window
-    // resize, sidebar drag, single↔grid flip, show/hide, and grid row-
-    // span changes without any push-based refit calls. Font-size changes
-    // don't change the body size and call _onBodyResize() explicitly.
-    this._pendingAttach = false;
-    this._revealRaf = 0;
-    this.ro = new ResizeObserver(() => this._onBodyResize());
-    this.ro.observe(this.body);
-
-    // WebGL renderer is dramatically faster than the default DOM
-    // renderer on older machines (VS Code uses the same approach).
-    // Load it lazily after open() and silently fall back to DOM if
-    // the GPU / driver doesn't support it. On context loss (e.g. the
-    // browser caps simultaneous WebGL contexts and kills ours when too
-    // many tiles exist), dispose the dead addon and try to re-attach;
-    // if re-attach fails, fall back to DOM and force a full repaint so
-    // we don't leave stale glyphs frozen on the canvas.
-    this._attachWebgl();
-
-    // Recover the renderer from the silent triggers that leave a stale
-    // backbuffer until the next resize: device-pixel-ratio changes
-    // (window dragged between displays with different scale, OS zoom)
-    // and visibility transitions (occlusion, GPU sleep). Both are cheap:
-    // clearTextureAtlas() rebuilds the glyph cache; term.refresh()
-    // forces a full repaint so the stale pixels are overwritten.
-    this._installRendererRecoveryListeners();
-
-    // Detect URLs in terminal output and route activation through
-    // the OS default browser. Hover underlines the URL; click (or
-    // ⌘-click when mouse reporting is active) follows it.
-    try {
-      this.term.loadAddon(new WebLinksAddon((event, uri) => {
-        if (uri) OpenURL(uri).catch(reportFailure('open link'));
-      }));
-    } catch (err) {
-      // Non-fatal; sessions still work without clickable links.
-    }
-
-    // When the running program enables mouse reporting (e.g. Claude,
-    // vim), xterm sends mousedown/mouseup to the PTY and cancels the
-    // event before the Linkifier can process it. Work around this by
-    // intercepting clicks on the xterm screen: if a recognized link
-    // is under the cursor, suppress the event so it doesn't reach the
-    // mouse protocol handler, letting the Linkifier's own handlers
-    // process it and call activate.
-    const screen = this.body.querySelector('.xterm-screen');
-    if (screen) {
-      screen.addEventListener('mousedown', (e) => {
-        const link = this.term._core?.linkifier?.currentLink;
-        if (link && link.link) {
-          this._pendingLink = link.link;
-          this._pendingLinkX = e.clientX;
-          this._pendingLinkY = e.clientY;
-          // Stop all other handlers — both the terminal's mouse-
-          // protocol handler and the Linkifier. We call activate
-          // manually on mouseup.
-          e.stopImmediatePropagation();
-        } else {
-          this._pendingLink = null;
-        }
-      }, { capture: true });
-      screen.addEventListener('mouseup', (e) => {
-        if (!this._pendingLink) return;
-        // Only treat as a click if the mouse barely moved (not a drag).
-        const dx = e.clientX - this._pendingLinkX;
-        const dy = e.clientY - this._pendingLinkY;
-        if (dx * dx + dy * dy < 25) {
-          e.stopImmediatePropagation();
-          this._pendingLink.activate(e, this._pendingLink.text);
-        }
-        this._pendingLink = null;
-      }, { capture: true });
-
-      // Click-to-position: send arrow keys to move the line-editor cursor
-      // to the clicked cell. Best-effort — only safe in the normal buffer
-      // with mouse reporting off; alt-buffer TUIs (vim/htop) own the screen.
-      screen.addEventListener('mousedown', (e) => {
-        if (e.button !== 0 || e.shiftKey || e.metaKey || e.ctrlKey || e.altKey) return;
-        this._pendingClickX = e.clientX;
-        this._pendingClickY = e.clientY;
-        this._pendingClick = true;
-      });
-      screen.addEventListener('mouseup', (e) => {
-        if (!this._pendingClick) return;
-        this._pendingClick = false;
-        const dx = e.clientX - this._pendingClickX;
-        const dy = e.clientY - this._pendingClickY;
-        if (dx * dx + dy * dy >= 25) return; // dragged → selection, leave it
-        const buf = this.term.buffer.active;
-        if (buf.type !== 'normal') return;
-        if (this.term.modes && this.term.modes.mouseTrackingMode && this.term.modes.mouseTrackingMode !== 'none') return;
-        const rect = screen.getBoundingClientRect();
-        const cellW = rect.width / this.term.cols;
-        const cellH = rect.height / this.term.rows;
-        if (!(cellW > 0) || !(cellH > 0)) return;
-        const col = Math.floor((e.clientX - rect.left) / cellW);
-        const row = Math.floor((e.clientY - rect.top) / cellH);
-        if (col < 0 || row < 0 || col >= this.term.cols || row >= this.term.rows) return;
-        // Only act when click is on the cursor's row — otherwise we'd send
-        // arrow-key spam that line editors partially consume and partially
-        // echo as literal "[D".
-        if (row !== buf.cursorY) return;
-        // Clamp the target column to the last non-space cell on this row,
-        // so clicking in the empty area past end-of-input does nothing.
-        const line = buf.getLine(buf.viewportY + row);
-        if (!line) return;
-        const text = line.translateToString(true);
-        const lastCol = text.replace(/\s+$/, '').length;
-        const target = Math.min(col, lastCol);
-        const delta = target - buf.cursorX;
-        if (delta === 0) return;
-        const seq = delta > 0 ? '\x1b[C'.repeat(delta) : '\x1b[D'.repeat(-delta);
-        this._writePty(seq);
-      });
-    }
-
-    // Single custom key handler. xterm.js keeps only ONE custom key event
-    // handler — a second attachCustomKeyEventHandler() call silently replaces
-    // the first — so every app-level binding must live here, not in a second
-    // registration.
-    this.term.attachCustomKeyEventHandler((e) => {
-      if (e.type !== 'keydown') return true;
-
-      // macOS Cmd+Backspace → Ctrl+U (kill to start of line). Browser doesn't
-      // translate this for us when xterm's helper-textarea is focused. Gated
-      // to mac so the Windows key on Linux/Windows can't accidentally fire it.
-      if (isMac && e.metaKey && !e.ctrlKey && !e.altKey && e.key === 'Backspace') {
-        e.preventDefault();
-        this._writePty('\x15');
-        return false;
-      }
-      // Shift+Enter → insert a newline in the agent's input instead of
-      // submitting. xterm sends a bare \r for Shift+Enter and drops the
-      // Shift, so Claude/Codex can't tell it from Enter and submit.
-      // NEWLINE_SEQ (Ctrl+J / \x0a) is the newline byte both agents accept
-      // with no terminal config. Plain Enter still submits. Cmd/Ctrl+Enter
-      // is intentionally not used here — it is the grid-project toggle, and
-      // the capture-phase window handler consumes it before xterm sees it.
-      if (isShiftEnter(e)) {
-        e.preventDefault();
-        this._writePty(NEWLINE_SEQ);
-        return false;
-      }
-      // App-level shortcuts that xterm would otherwise translate into a
-      // control sequence and forward to the PTY (where the shell beeps because
-      // the binding is meaningless). Returning false tells xterm to ignore the
-      // event; it still bubbles to the window-level keydown handler that runs
-      // the actual shortcut. Ctrl+` is intentionally Ctrl-only on every
-      // platform (mirrors VS Code; macOS already uses ⌘` for window cycling).
-      if (e.ctrlKey && !e.metaKey && e.code === 'Backquote') {
-        return false;
-      }
-
-      // Ctrl+Shift copy/paste/select-all. Required because when an inner
-      // program (Claude CLI, vim) enables DEC mouse tracking, xterm.js
-      // forwards mouse events to the PTY instead of using them for text
-      // selection — leaving no way to copy output. Ctrl+Shift+A selects the
-      // full scrollback, Ctrl+Shift+C copies the current selection,
-      // Ctrl+Shift+V pastes from the system clipboard.
-      if (!e.ctrlKey || !e.shiftKey || e.altKey || e.metaKey) return true;
-      const key = e.key.toLowerCase();
-      if (key === 'c') {
-        const sel = this.term.getSelection();
-        // SetClipboardText (Go-side via atotto/clipboard) rather than
-        // wails runtime.ClipboardSetText — the latter is broken on
-        // Windows (non-STA goroutine, OpenClipboard fails silently).
-        if (sel) SetClipboardText(sel).catch(reportFailure('copy'));
-        return false;
-      }
-      if (key === 'v') {
-        ClipboardGetText().then((text) => {
-          if (text) this._writePty(text);
-        }).catch(reportFailure('paste'));
-        return false;
-      }
-      if (key === 'a') {
-        this.term.selectAll();
-        return false;
-      }
-      return true;
-    });
-
-    // Visual focus (.term-focused) and keyboard focus (xterm's
-    // helper-textarea) are reconciled atomically by setFocusedTile(id),
-    // which is the sole writer of .term-focused. Driving the class off
-    // browser focusin/focusout events used to race with DOM churn during
-    // view transitions (single ↔ grid, renderGrid's appendChild reorder,
-    // xterm.open mounting new helper-textareas), leaving a tile visually
-    // focused while keystrokes went nowhere. Visual focus is now a pure
-    // projection of state.activeId, gated by whether a modal/rename
-    // owns the keyboard — they can't drift.
-
-    this.attached = false;
-    // needsReattach is set by pty:disconnect when our attach connection
-    // drops (e.g. Restart Session closes the daemon-side PTY). The next
-    // session:event(updated, alive=true) consumes the flag and triggers
-    // ensureAttached so the new PTY's stream resumes without the user
-    // having to switch sessions and back.
-    this.needsReattach = false;
-
-    // Track the OSC-set window title from the running TUI (vim, htop,
-    // claude code, etc.) so the app title bar can show it after the
-    // session name when this session is active.
-    this.termTitle = '';
-    this.term.onTitleChange((title) => {
-      this.termTitle = title || '';
-      this._renderTermTitle();
-      if (state.activeId === this.info.id) updateAppTitle();
-    });
-
-    this._writePty = (data) => {
-      const bytes = new TextEncoder().encode(data);
-      let bin = '';
-      for (let i = 0; i < bytes.length; i++) bin += String.fromCharCode(bytes[i]);
-      // Intentionally no reportFailure: this fires per keystroke, so a
-      // dead daemon would flood the status bar with one error per key.
-      // The disconnect itself is surfaced once ("control disconnected").
-      WriteStdin(this.info.id, btoa(bin));
-    };
-    this.term.onData((data) => this._writePty(data));
-
-    // Click anywhere on the tile (header or body) selects this session.
-    this.host.addEventListener('mousedown', () => {
-      if (state.activeId !== this.info.id) {
-        setActive(this.info.id);
-        updateSidebarSelection();
-        if (state.view === 'single') {
-          // Switch terms in single mode; in grid mode every tile is
-          // already visible so there's nothing else to do.
-          showSingle(this.info.id);
-        } else {
-          renderGrid();
-        }
-      } else {
-        // Reclick on the active tile — still clears any leftover
-        // attention indicator.
-        clearAttention(this.info.id);
-      }
-    });
-
-    // Double-click the tile name to rename inline (same affordance
-    // as the sidebar). The header's mousedown selects the tile;
-    // dblclick on the name then opens the editor.
-    this.tileName.addEventListener('dblclick', (e) => {
-      e.stopPropagation();
-      this._beginRename();
-    });
-
-    // BEL on a non-focused session marks it as needing attention and
-    // fires a desktop notification. xterm.js v5 exposes onBell.
-    this.term.onBell(() => {
-      onSessionBell(this.info);
-    });
-
-    // Dead-session overlay: hidden until the underlying process exits
-    // (Alive transitions true→false). Centered card with primary
-    // "Close session" (Enter) and secondary "Dismiss" (Escape).
-    this.deadOverlay = document.createElement('div');
-    this.deadOverlay.className = 'dead-overlay';
-    this.deadOverlay.setAttribute('role', 'alertdialog');
-    this.deadOverlay.setAttribute('aria-label', 'Session ended');
-    this.deadOverlay.hidden = true;
-    const card = document.createElement('div');
-    card.className = 'dead-card';
-    const title = document.createElement('div');
-    title.className = 'dead-title';
-    title.textContent = 'Session ended';
-    const subtitle = document.createElement('div');
-    subtitle.className = 'dead-subtitle';
-    subtitle.textContent = 'The process running in this session has exited.';
-    const buttons = document.createElement('div');
-    buttons.className = 'dead-buttons';
-    this.deadCloseBtn = document.createElement('button');
-    this.deadCloseBtn.className = 'dead-btn primary';
-    this.deadCloseBtn.textContent = 'Close session';
-    this.deadCloseBtn.addEventListener('click', (e) => {
-      e.stopPropagation();
-      this._closeDead();
-    });
-    this.deadDismissBtn = document.createElement('button');
-    this.deadDismissBtn.className = 'dead-btn secondary';
-    this.deadDismissBtn.textContent = 'Dismiss';
-    this.deadDismissBtn.addEventListener('click', (e) => {
-      e.stopPropagation();
-      this._dismissDead();
-    });
-    buttons.append(this.deadCloseBtn, this.deadDismissBtn);
-    card.append(title, subtitle, buttons);
-    this.deadOverlay.append(card);
-    this.host.append(this.deadOverlay);
-    this.deadOverlayShown = false;
-
-    // Take over wheel handling. xterm's default wheel→lines math
-    // honors raw deltaY, which on macOS trackpads with momentum
-    // pumps in events of 200–400px each — enough to fly past a
-    // 5000-line scrollback in a single swipe. We cap each event to
-    // a sane line count so the user can actually read history.
-    // Capture phase + stopPropagation prevents xterm's own handler
-    // from firing.
-    const linesPerPixel = 1 / 14; // ~one line per ~14 px of delta
-    const maxLinesPerEvent = 8;   // about half a screen on a small tile
-    this.host.addEventListener('wheel', (e) => {
-      e.preventDefault();
-      e.stopPropagation();
-      let lines = Math.round(e.deltaY * linesPerPixel);
-      if (lines === 0) {
-        // Sub-pixel events — preserve direction so a slow scroll
-        // still moves at least one line.
-        lines = e.deltaY > 0 ? 1 : e.deltaY < 0 ? -1 : 0;
-      }
-      if (lines > maxLinesPerEvent) lines = maxLinesPerEvent;
-      if (lines < -maxLinesPerEvent) lines = -maxLinesPerEvent;
-      if (lines !== 0) this.term.scrollLines(lines);
-    }, { capture: true, passive: false });
-  }
-
-  _attachWebgl() {
-    // Build / re-build the WebGL addon. Called at init and on
-    // context-loss recovery. After a fresh attach the renderer's atlas
-    // is empty, so force a full repaint to overwrite whatever stale
-    // pixels were left behind by the lost context.
-    try {
-      const webgl = new WebglAddon();
-      webgl.onContextLoss(() => this._onWebglContextLoss());
-      this.term.loadAddon(webgl);
-      this.webgl = webgl;
-      try { this.term.refresh(0, this.term.rows - 1); } catch {}
-      return true;
-    } catch {
-      this.webgl = null;
-      return false;
-    }
-  }
-
-  _onWebglContextLoss() {
-    // The current addon's context died (commonly: too many WebGL
-    // contexts process-wide). Recovery logic — dispose, reattach, fall
-    // back to refresh — lives in lib/renderer-recovery.js so it can be
-    // unit-tested without xterm or a real WebGL context.
-    const dead = this.webgl;
-    this.webgl = null;
-    recoverFromContextLoss({
-      dispose: () => dead?.dispose(),
-      reattach: () => this._attachWebgl(),
-      refresh: () => this.term.refresh(0, this.term.rows - 1),
-    });
-  }
-
-  _installRendererRecoveryListeners() {
-    // Clear the glyph atlas and force a full repaint. Cheap; safe to
-    // call when no WebGL addon is loaded (DOM renderer ignores the
-    // atlas hint and still benefits from the refresh).
-    this._refreshRenderer = () => {
-      try { this.webgl?.clearTextureAtlas(); } catch {}
-      try { this.term.refresh(0, this.term.rows - 1); } catch {}
-    };
-
-    // DPR change: move-to-different-display or OS zoom. A
-    // `(resolution: Xdppx)` MQL only fires `change` on the single
-    // transition away from X, so the helper rebinds against the new
-    // DPR inside each handler — feature-detected and self-teardown so
-    // it's safe on Chromium-CEF builds that don't expose matchMedia.
-    this._dprWatcher = bindDprWatcher({
-      matchMedia: (q) => window.matchMedia(q),
-      getDpr: () => window.devicePixelRatio || 1,
-      onChange: () => this._refreshRenderer(),
-    });
-
-    // Visibility transitions: occlusion / GPU sleep can invalidate the
-    // backbuffer without firing context-loss. Repaint on return.
-    this._onVisibility = () => {
-      if (shouldRefreshOnVisibility(document.visibilityState)) this._refreshRenderer();
-    };
-    document.addEventListener('visibilitychange', this._onVisibility);
-  }
-
-  setInfo(info) {
-    this.info = info;
-    this.host.style.setProperty('--session-color', info.color || '#888');
-    this.tileName.textContent = info.name;
-    this.header.setAttribute('aria-label', `Session ${info.name}`);
-    const wtBranch = info.worktreeBranch ?? info.worktree_branch;
-    if (wtBranch) {
-      this.tileWorktree.style.display = '';
-      this.tileWorktree.title = `Worktree: ${wtBranch}`;
-    } else {
-      this.tileWorktree.style.display = 'none';
-      this.tileWorktree.title = '';
-    }
-    this._renderTermTitle();
-  }
-
-  _renderTermTitle() {
-    // Hide the slot when the TUI hasn't set a title or it just echoes
-    // the session name (avoids "foo — foo").
-    const t = this.termTitle || '';
-    if (!t || t === this.info.name) {
-      this.tileTermTitle.textContent = '';
-      this.tileTermTitle.style.display = 'none';
-    } else {
-      this.tileTermTitle.textContent = t;
-      this.tileTermTitle.title = t;
-      this.tileTermTitle.style.display = '';
-    }
-  }
-
-  // _beginRename hides the tile-name span, drops an input next to
-  // it, and calls UpdateSession on Enter / blur. Escape cancels.
-  // The next session:event(updated) calls setInfo which refreshes
-  // tileName.textContent; we just need to put the span back in DOM.
-  _beginRename() {
-    if (this._renameInput) return; // already editing
-    const input = document.createElement('input');
-    input.type = 'text';
-    input.className = 'tile-name-input';
-    input.value = this.info.name;
-    this._renameInput = input;
-    this.tileName.style.display = 'none';
-    this.tileName.parentNode.insertBefore(input, this.tileName);
-    // Drop the visual focus border before stealing keyboard focus —
-    // setFocusedTile is the only writer of .term-focused, so without
-    // this the border would linger while the rename input owns input.
-    setFocusedTile(null);
-    input.focus();
-    input.select();
-    let done = false;
-    const finish = (commit) => {
-      if (done) return;
-      done = true;
-      const next = input.value.trim();
-      input.remove();
-      this._renameInput = null;
-      this.tileName.style.display = '';
-      if (commit && next && next !== this.info.name) {
-        UpdateSession(this.info.id, next, '', -1).catch(reportFailure('rename'));
-      }
-      refocusActiveTerm();
-    };
-    // ONE capture-phase listener handles Enter/Escape AND shields the
-    // input from xterm / global hotkey handlers. It must be a single
-    // listener: stopPropagation() from a capture listener at the target
-    // also cancels the target's own bubble-phase listeners (DOM dispatch
-    // skips the bubble invocation once the flag is set), so a separate
-    // bubble-phase Enter handler would never run — Enter/Escape were
-    // dead and renames only ever committed via blur.
-    input.addEventListener('keydown', (e) => {
-      e.stopPropagation();
-      if (e.key === 'Enter') finish(true);
-      else if (e.key === 'Escape') finish(false);
-    }, { capture: true });
-    input.addEventListener('blur', () => finish(true));
-  }
-
-  setProject(name, color) {
-    this.tileProject.textContent = name || '';
-    this.host.style.setProperty('--project-color', color || '#888');
-  }
-
-  show() {
-    // Becoming visible flips display from none → block. fit.fit()
-    // updates xterm's cols/rows synchronously, but the WebGL renderer
-    // schedules the *canvas pixel resize* on rAF — so for one frame
-    // the stale (grid-cell-sized, or last-zoom-sized) canvas is CSS-
-    // stretched into the new body box, producing a huge-text flash.
-    // Gate paint with visibility:hidden across the rAF: layout still
-    // runs (fit measures the real body box), only the pixel paint is
-    // suppressed until after the renderer has caught up.
-    this.body.style.visibility = 'hidden';
-    this.host.classList.add('visible');
-    void this.body.clientWidth;
-    this._onBodyResize();
-    if (this._revealRaf) cancelAnimationFrame(this._revealRaf);
-    this._revealRaf = requestAnimationFrame(() => {
-      this._revealRaf = 0;
-      this.body.style.visibility = '';
-    });
-  }
-
-  hide() {
-    this.host.classList.remove('visible');
-    // Cancel any in-flight reveal so the next show() starts from a
-    // known-good state if the user switches away during the rAF gate.
-    if (this._revealRaf) {
-      cancelAnimationFrame(this._revealRaf);
-      this._revealRaf = 0;
-    }
-    this.body.style.visibility = '';
-  }
-
-  // _onBodyResize is the single resize entry point. ResizeObserver
-  // delivers the call post-layout, so fit.fit() reads correct dims
-  // — no rAF dance needed. Font-size changes call this explicitly
-  // (the body box doesn't change, so RO won't fire on its own).
-  _onBodyResize() {
-    // RO can fire with a zero box when the host is display:none (tile
-    // not .visible, or hidden because outside the grid scope). fit.fit()
-    // on a zero-size body produces garbage dims — skip until visible.
-    if (this.body.clientWidth === 0 || this.body.clientHeight === 0) return;
-
-    // First-time visibility for a deferred attach: hand off to
-    // ensureAttached, which does its own fit.fit() before OpenSession.
-    if (this._pendingAttach) {
-      this._pendingAttach = false;
-      this.ensureAttached();
-      return;
-    }
-
-    // Preserve "viewport pinned to bottom" across the resize. xterm's
-    // own resize doesn't auto-snap to bottom after reflow; without this,
-    // a user scrolled to the latest line would land mid-history.
-    // Tolerance: codex (and similar TUIs) sometimes leave the viewport
-    // a line or two short of the bottom; treat that as "at bottom" so
-    // resize doesn't strand the user mid-history. Anything further up
-    // is deliberate scrollback — leave it alone.
-    const STICKY_BOTTOM_LINES = 2;
-    const buf = this.term.buffer.active;
-    const wasAtBottom = buf ? (buf.baseY - buf.viewportY) <= STICKY_BOTTOM_LINES : true;
-    // Swallow throw and continue: a transient FitAddon error (e.g. a
-    // race against teardown) shouldn't drop the daemon-side resize.
-    const prevCols = this.term.cols;
-    try { this.fit.fit(); } catch { /* keep going with last-known dims */ }
-    if (this.attached) {
-      // Intentionally no reportFailure: resize fires continuously during
-      // window/sidebar drags, so a dead daemon would flood the status
-      // bar. The disconnect is surfaced once ("control disconnected").
-      ResizeSession(this.info.id, this.term.cols, this.term.rows);
-    }
-    if (wasAtBottom) this.term.scrollToBottom();
-
-    // If the column count changed materially relative to the
-    // *baseline* (the cols active at the last replay, or initial
-    // attach if no replay has fired yet), xterm's scrollback is
-    // stale — its rendered rows were baked at the old width and
-    // xterm.js does not reflow history on resize. Ask the daemon to
-    // re-stream the raw byte ring; the EventScrollbackReplayBegin
-    // handler will term.reset() before the bytes arrive, and the
-    // daemon serializes the replay against live fanout so nothing
-    // interleaves.
-    //
-    // Comparing against a baseline (rather than the just-previous
-    // measurement) means a 80→84→83 sequence still triggers a replay
-    // — the final width is 3 cols off the baseline, even though
-    // neither single step crosses the threshold. We also unconditionally
-    // clear any pending timer on every resize, then re-arm only if
-    // the *current* delta still warrants a replay; otherwise an old
-    // measurement that briefly crossed the threshold would leave a
-    // stale timer armed.
-    if (this._replayBaselineCols === undefined) {
-      this._replayBaselineCols = prevCols || this.term.cols;
-    }
-    if (this._replayTimer) {
-      clearTimeout(this._replayTimer);
-      this._replayTimer = 0;
-      // Cancel-without-rearm must also clear any stale wants-bottom
-      // intent — otherwise a `false` captured at the previous resize
-      // outlives its replay and suppresses the bottom-snap on the
-      // next replay-done from any source (re-attach, daemon-initiated
-      // atomic replay on subscribe, etc.).
-      delete this._replayWantsBottom;
-    }
-    if (scrollTrace.rec.enabled) {
-      scrollTrace.rec('resize', {
-        id: this.info.id, prevCols, cols: this.term.cols,
-        baseline: this._replayBaselineCols, wasAtBottom,
-      });
-    }
-    if (this.attached && shouldRequestReplay(this._replayBaselineCols, this.term.cols)) {
-      // Carry the user's pre-resize "at bottom?" intent through to the
-      // scrollback_replay_done handler. If the user was actively reading
-      // scrollback (wasAtBottom === false), the replay must not yank
-      // them back to the bottom on completion.
-      this._replayWantsBottom = wasAtBottom;
-      this._replayTimer = setTimeout(() => {
-        this._replayTimer = 0;
-        this._replayBaselineCols = this.term.cols;
-        if (scrollTrace.rec.enabled) {
-          scrollTrace.rec('replay-request', { id: this.info.id, cols: this.term.cols });
-        }
-        RequestScrollbackReplay(this.info.id).catch(() => { /* attach may have closed */ });
-      }, REPLAY_DEBOUNCE_MS);
-    }
-  }
-
-  // rebaselineReplayCols resets _replayBaselineCols to the current
-  // term.cols and clears any pending replay timer. Use this when grid
-  // geometry changes for a reason that is NOT a user-driven window
-  // resize (first-attach in grid; minimize/restore reflowing the
-  // remaining tiles). Those reflows shrink/widen the tile but the
-  // scrollback was already written at the new width once xterm
-  // re-fitted, so triggering shouldRequestReplay would be spurious and
-  // visibly drops or duplicates content. Pure window resize is handled
-  // by the threshold path in _onBodyResize and must not call this.
-  rebaselineReplayCols(_reason) {
-    applyRebaseline(this);
-  }
-
-  async ensureAttached() {
-    if (this.attached) return;
-    // Don't attempt to attach to a session known to be dead — the daemon
-    // will refuse. Show the dead overlay with the error reason instead.
-    if (state.aliveById.get(this.info.id) === false) {
-      this.setDead(true, this.info.last_error || 'The process failed to start.');
-      return;
-    }
-    // If the host is still display:none, the body has no box yet and
-    // fit.fit() would measure 0×0. Defer until ResizeObserver fires
-    // with a real size — _onBodyResize will re-enter ensureAttached.
-    if (this.body.clientWidth === 0 || this.body.clientHeight === 0) {
-      this._pendingAttach = true;
-      return;
-    }
-    this.fit.fit();
-    try {
-      await OpenSession(this.info.id, this.term.cols, this.term.rows);
-      this.attached = true;
-      // Anchor the replay baseline to the actual fitted cols for this
-      // tile. Without this, a later _onBodyResize would initialize the
-      // baseline from a stale xterm default (80) while term.cols is the
-      // real grid-cell width — the next resize crosses the threshold
-      // and fires a spurious scrollback replay on first grid entry.
-      this.rebaselineReplayCols('first-attach');
-    } catch (err) {
-      this.term.write(`\r\n\x1b[31m[attach failed: ${err}]\x1b[0m\r\n`);
-    }
-  }
-
-  writeData(b64) {
-    const bin = atob(b64);
-    const bytes = new Uint8Array(bin.length);
-    for (let i = 0; i < bin.length; i++) bytes[i] = bin.charCodeAt(i);
-    this.term.write(this.decoder.decode(bytes, { stream: true }));
-  }
-
-  destroy() {
-    // Intentionally silent: destroy() tears down a session that's already
-    // gone; a failed CloseAttach has nothing for the user to act on.
-    CloseAttach(this.info.id).catch(() => {});
-    if (this._revealRaf) cancelAnimationFrame(this._revealRaf);
-    this.ro.disconnect();
-    if (this._dprWatcher) {
-      try { this._dprWatcher.teardown(); } catch {}
-      this._dprWatcher = null;
-    }
-    if (this._onVisibility) {
-      document.removeEventListener('visibilitychange', this._onVisibility);
-    }
-    // Release the GL context proactively so a many-tile session doesn't
-    // sit on it until GC and push another tile over the browser cap.
-    try { this.webgl?.dispose(); } catch {}
-    this.webgl = null;
-    this.term.dispose();
-    this.host.remove();
-  }
-
-  setDead(isDead, reason) {
-    this.deadOverlayShown = isDead;
-    this.deadOverlay.hidden = !isDead;
-    this.host.classList.toggle('dead', isDead);
-    if (isDead) {
-      const subtitle = this.deadOverlay.querySelector('.dead-subtitle');
-      if (subtitle && reason) {
-        subtitle.textContent = reason;
-      }
-      // Defer focus so it lands after the visibility flip and after
-      // any pending blur from the dying xterm.
-      setTimeout(() => {
-        if (this.deadOverlayShown) this.deadCloseBtn.focus();
-      }, 0);
-    }
-  }
-
-  _closeDead() {
-    KillSession(this.info.id, true).catch(reportFailure('close'));
-  }
-
-  _dismissDead() {
-    state.dismissedDead.add(this.info.id);
-    this.setDead(false);
-    refocusActiveTerm();
-  }
-}
-
-function applyFontSize() {
-  for (const st of state.terms.values()) {
-    st.term.options.fontSize = state.fontSize;
-    // Body box doesn't change on font-size change, so ResizeObserver
-    // won't fire — call the resize handler explicitly so fit.fit()
-    // recomputes (cols, rows) from new char metrics.
-    st._onBodyResize();
-  }
-  localStorage.setItem('hive.fontSize', String(state.fontSize));
-}
-
-function bumpFontSize(delta) {
-  const next = clampFont(state.fontSize + delta);
-  if (next === state.fontSize) return;
-  state.fontSize = next;
-  applyFontSize();
-  // flashStatus (not setStatus): per-action feedback must auto-revert,
-  // not overwrite the persistent slot ("control disconnected", session
-  // name) until the next nav event.
-  flashStatus(`font ${state.fontSize}px`);
-}
-
-function resetFontSize() {
-  state.fontSize = DEFAULT_FONT_SIZE;
-  applyFontSize();
-  flashStatus(`font ${state.fontSize}px`);
-}
-
-function ensureTerm(info) {
-  let st = state.terms.get(info.id);
-  if (!st) {
-    st = new SessionTerm(info);
-    state.terms.set(info.id, st);
-  } else {
-    st.setInfo(info);
-  }
-  const proj = state.projects.find((p) => p.id === (info.projectId ?? info.project_id));
-  st.setProject(proj?.name ?? '', proj?.color ?? '');
-  return st;
-}
-
-function setActive(id) {
-  if (id) {
-    state.attention.delete(id);
-    state.terms.get(id)?.host.classList.remove('attention');
-    const s = state.sessions.find((x) => x.id === id);
-    const pid = s?.projectId ?? s?.project_id;
-    if (pid) state.currentProjectId = pid;
-  }
-  state.activeId = id;
-  // Schedule focus after the next paint so any DOM reorder / visibility
-  // change from renderGrid / showSingle has settled. xterm.focus()
-  // moves focus to its hidden textarea; that fires .onFocus, which
-  // adds .term-focused to the host (the source of truth for the
-  // visual focus border).
-  if (id) focusActiveTerm();
-}
-
-// setFocusedTile is the SOLE writer of .term-focused. It reconciles
-// visual focus and keyboard focus atomically — in the same rAF tick —
-// so they can never drift. Visual focus is a pure projection of
-// state.activeId gated by whether a modal/rename owns the keyboard.
-//
-// Pass state.activeId to focus the active session. Pass null to drop
-// the visual focus everywhere (e.g. when opening the launcher or a
-// rename input). Every state transition that could change which tile
-// should be focused (setActive, setView, renderGrid, modal open/close,
-// rename open/close, dialog close, OS fullscreen toggle, …) MUST end
-// by calling setFocusedTile(...).
-//
-// Previously visual focus was event-driven from focusin/focusout on
-// each .term-host, and keyboard focus was driven separately by
-// focusActiveTerm()'s ta.focus() call. During view transitions (most
-// visibly single → grid: renderGrid's appendChild reorder and the
-// helper-textarea mounted by xterm.open() for newly-materialized
-// tiles), the two could end up on different tiles. The user would
-// see a session lit up while keystrokes went nowhere. Single writer
-// makes that impossible: the class is added only here, in the same
-// rAF as the helper-textarea focus.
-// Transient focus guard for the post-view-switch settle window.
-//
-// Switching to grid reparents the active tile (renderGrid's appendChild
-// reorder) and triggers async ResizeObserver → fit → WebGL resize on the
-// newly-visible neighbour tiles. Both momentarily blur the active
-// helper-textarea to <body>. A keystroke typed in that sub-frame gap lands
-// on <body> and is silently lost — observed as "ello"/"o" instead of
-// "hello" right after ⌘⇧G (a real keystroke-loss bug, and the cause of the
-// flaky focus E2E). applyFocus's rAF retry re-focuses, but only on the NEXT
-// frame, too late for a char already dropped.
-//
-// This document-level capture guard re-focuses SYNCHRONOUSLY the instant the
-// guarded textarea blurs to <body>, so focus is back before the next
-// keystroke's event-loop turn. It is armed only for a short window after a
-// real-tile focus request and only acts while that tile is still active and
-// no modal/rename legitimately owns the keyboard — so it never traps focus.
-let _focusGuard = null; // { id, until } | null
-
-function armFocusGuard(id) {
-  _focusGuard = { id, until: performance.now() + 500 };
-}
-
-document.addEventListener(
-  'focusout',
-  (e) => {
-    const g = _focusGuard;
-    if (!g) return;
-    if (performance.now() > g.until) { _focusGuard = null; return; }
-    if (state.activeId !== g.id) return; // active tile changed → let it go
-    if (focusSnapshot(g.id).modalOpen) return; // modal/rename owns keyboard
-    const st = state.terms.get(g.id);
-    if (!st) return;
-    const ta = st.host.querySelector('.xterm-helper-textarea');
-    if (!ta || e.target !== ta) return; // only when OUR textarea blurs
-    // Only reclaim a transient blur to nothing/<body>; never override the
-    // user intentionally focusing another control.
-    const dest = e.relatedTarget;
-    if (dest && dest !== document.body) return;
-    ta.focus();
-  },
-  true,
-);
-
-function setFocusedTile(id) {
-  // First decision: synchronous, before any rAF. If we already know we
-  // should clear, do it immediately so a modal/null transition can't be
-  // overtaken by a stale in-flight focus rAF.
-  const snap = focusSnapshot(id);
-  if (snap.modalOpen || id == null || !state.terms.get(id)) {
-    _focusGuard = null;
-    sweepFocusBorder();
-    return;
-  }
-  // Arm the synchronous blur guard for the settle window, then schedule
-  // the focus drive after the next paint so any in-flight DOM transition
-  // (showSingle / renderGrid / appendChild / xterm.open) settles before we
-  // read activeElement and move focus.
-  armFocusGuard(id);
-  requestAnimationFrame(() => applyFocus(id, /*attempt=*/0));
-}
-
-function applyFocus(id, attempt) {
-  const st = state.terms.get(id);
-  if (!st) { sweepFocusBorder(); return; }
-  const action = decideFocusAction(focusSnapshot(id));
-  if (action.kind === ACTION_CLEAR || action.kind === ACTION_PRESERVE) {
-    sweepFocusBorder();
-    return;
-  }
-  // Atomic reconcile: sweep + add + focus.
-  for (const el of document.querySelectorAll('.term-host.term-focused')) {
-    if (el !== st.host) el.classList.remove('term-focused');
-  }
-  st.host.classList.add('term-focused');
-  // Drive browser focus to the DOM helper-textarea. xterm's
-  // term.focus() early-returns on a stale-true _focused flag (#159);
-  // after this transition the flag is stale-false because the
-  // synchronous display:none flip during renderGrid's parent class
-  // swap (single → grid) fires focusout. ta.focus() drives the real
-  // event; the follow-up term.focus() resyncs xterm's internal state.
-  const ta = st.host.querySelector('.xterm-helper-textarea');
-  // Only drive focus when it has actually drifted off the target
-  // textarea. Re-focusing an already-focused xterm helper-textarea is
-  // NOT a harmless no-op: it clears the textarea's pending input mid-
-  // keystroke, so a character typed during the post-grid-switch retry
-  // window is dropped before xterm's input event emits it (observed as
-  // "ello" / "o" instead of "hello"). Because several setFocusedTile
-  // calls fire during a grid switch, multiple retry chains overlap and
-  // hammer focus() every frame for ~300ms; guarding on real drift keeps
-  // the #159/#181/#186 drift-correction while ending the keystroke loss.
-  if (ta && document.activeElement !== ta) {
-    ta.focus();
-    if (typeof st.term?.focus === 'function') st.term.focus();
-  }
-  // Schedule a verification rAF *next frame* (not this one — focus()
-  // just fired and synchronously updated activeElement, so an in-tick
-  // check would trivially pass and miss the real failure mode):
-  // post-renderGrid side-effects (ResizeObserver → fit → WebGL canvas
-  // resize on newly-visible neighbour tiles) can synchronously fire
-  // focusout ~10ms later. If activeElement has drifted off `ta` by
-  // then, re-focus. Cap retries so a genuine modal-takeover or rename
-  // doesn't busy-loop.
-  // Poll for several frames. A single rAF check is insufficient
-  // because post-renderGrid side-effects (ResizeObserver → fit →
-  // WebGL canvas resize on neighbour tiles) can fire focusout AFTER
-  // the rAF batch completes — the disturbance arrives one event-loop
-  // turn later than the verify. We watch for FOCUS_MAX_RETRIES frames
-  // and re-focus whenever activeElement drifts off `ta`. Polling is
-  // bounded and idempotent (re-focusing an already-focused element is
-  // a no-op).
-  const FOCUS_MAX_RETRIES = 8;
-  if (ta && attempt < FOCUS_MAX_RETRIES) {
-    requestAnimationFrame(() => {
-      const verifyAction = decideFocusAction(focusSnapshot(id));
-      if (verifyAction.kind !== ACTION_FOCUS) return; // a modal / rename took over
-      applyFocus(id, attempt + 1);
-    });
-  }
-  // Optional dev-mode assertion: two rAFs later, the visual focus
-  // and the keyboard focus should agree. Console-warn on drift so
-  // future variants of #159/#181/#186 are caught in QA.
-  if (debugFocusEnabled() && attempt === 0) scheduleFocusConsistencyCheck(id);
-}
-
-function sweepFocusBorder() {
-  for (const el of document.querySelectorAll('.term-host.term-focused')) {
-    el.classList.remove('term-focused');
-  }
-}
-
-function focusSnapshot(id) {
-  const ae = document.activeElement;
-  return {
-    id,
-    modalOpen: anyModalOpen(),
-    activeTag: ae ? ae.tagName : '',
-    activeClasses: ae ? ae.classList : '',
-    knownTermIds: state.terms,
-  };
-}
-
-function debugFocusEnabled() {
-  try { return localStorage.getItem('hive.debug') === '1'; } catch { return false; }
-}
-
-function scheduleFocusConsistencyCheck(id) {
-  requestAnimationFrame(() => requestAnimationFrame(() => {
-    const st = state.terms.get(id);
-    if (!st) return;
-    const ta = st.host.querySelector('.xterm-helper-textarea');
-    const ae = document.activeElement;
-    const focusedHost = ae ? ae.closest('.term-host') : null;
-    if (focusedHost !== st.host || ae !== ta) {
-      // eslint-disable-next-line no-console
-      console.warn('[focus] inconsistent state', {
-        view: state.view,
-        activeId: state.activeId,
-        wantId: id,
-        aeTag: ae ? ae.tagName : null,
-        aeClass: ae ? ae.className : null,
-        focusedHostMatches: focusedHost === st.host,
-      });
-    }
-  }));
-}
-
-// focusActiveTerm / refocusActiveTerm are thin wrappers retained so
-// every existing callsite (and any third-party readers of the code)
-// keeps working. Both reduce to setFocusedTile(state.activeId): the
-// gate inside setFocusedTile decides whether to apply or clear.
-function focusActiveTerm() {
-  setFocusedTile(state.activeId);
-}
-
-function refocusActiveTerm() {
-  setFocusedTile(state.activeId);
-}
-
-// gridSpatialMove moves the active tile in the given direction.
-// Uses cellMap to honor row-spanned tiles: e.g. with 3 sessions in a
-// 2x2 grid the bottom-right cell is absorbed by tile 1, so pressing
-// "right" from tile 2 lands on tile 1 instead of doing nothing.
-// ---------- command palette ----------
+import { ensureTerm, bumpFontSize, resetFontSize } from './app/session-term.js';
+import {
+  setActive, setFocusedTile, focusActiveTerm, refocusActiveTerm, initFocus,
+} from './app/focus.js';
+
+// ---------- command palette table ----------
 
 // Shortcut strings come from lib/shortcuts.js so the palette and the
 // ⌘/ help overlay can't drift from each other.
@@ -1146,9 +77,12 @@ const paletteCommands = [
   })),
 ].map((c) => ({ ...c, shortcut: PALETTE_KEYS[c.id] ?? '' }));
 
-// Modal wiring: focus-pipeline callbacks are injected so the modal
-// modules never import the pipeline (main.js owns that coupling), and
-// each modal registers itself with the registry focusSnapshot reads.
+// ---------- wiring ----------
+
+// Cross-module callbacks are injected here so the modules stay
+// acyclic: modals/sidebar/view/keyboard/events never import the focus
+// pipeline or SessionTerm directly. Each modal also registers itself
+// with the registry focusSnapshot reads.
 initLauncher({ setFocusedTile, refocusActiveTerm });
 initProjectEditor({ setFocusedTile, refocusActiveTerm });
 initCommandPalette({ commands: paletteCommands, focusActiveTerm });
@@ -1157,6 +91,7 @@ initSidebar({ switchTo, switchToProject, confirmAndDeleteProject, renderEmptySta
 initBanners();
 initView({ ensureTerm, setActive, focusActiveTerm, scrollTrace });
 initKeyboard({ bumpFontSize, resetFontSize, focusActiveTerm });
+initFocus({ ensureTerm });
 wireDaemonEvents({
   switchTo, renderMinimizedTray, updateAppTitle, focusActiveTerm,
   refocusActiveTerm, isDaemonRestarting, scrollTrace,
@@ -1168,8 +103,6 @@ wireDaemonEvents({
 const footerHintsEl = document.getElementById('sidebar-hints');
 if (footerHintsEl) footerHintsEl.textContent = footerHints({ isMac });
 
-// moveActiveSession walks the (project_order, session_order) list.
-// reorder=true moves the session within its project only.
 // ---------- sidebar resize ----------
 //
 // Drag the right edge of the sidebar to resize. Width persists across
